models/skill.py
from extensions import db
from datetime import datetime
from sqlalchemy import CheckConstraint

class Skill(db.Model):
    __tablename__ = 'skills'
    id = db.Column(db.Integer, primary_key=True)
    skill_name = db.Column(db.String(100), unique=True, nullable=False)
    category = db.Column(db.String(50))
    description = db.Column(db.Text)
    created_at = db.Column(db.DateTime, default=datetime.utcnow)

class CandidateSkill(db.Model):
    __tablename__ = 'candidate_skills'
    id = db.Column(db.Integer, primary_key=True)
    candidate_id = db.Column(db.Integer, db.ForeignKey('candidate_profiles.id'), nullable=False)
    skill_id = db.Column(db.Integer, db.ForeignKey('skills.id'), nullable=False)
    proficiency_level = db.Column(db.String(20), default='Intermediate')
    years_experience = db.Column(db.Integer, default=0)
    
    __table_args__ = (
        CheckConstraint("proficiency_level IN ('Beginner', 'Intermediate', 'Advanced', 'Expert')", name='proficiency_level_check'),
    )
